Adding many metrics should similarly happen all at once.

+++ src/modules/copied/actions.js
@@ -68,9 +68,12 @@ export function paste(spaceId){
     )
 
     const existingMetrics = spaceMetrics.filter(m => isWithinRegion(m.location, pasteRegion))
-    dispatch(metricActions.removeMetrics(existingMetrics.map(m => m.id)))
-
-    newMetrics.forEach((newMetric, i) => {dispatch({type: 'ADD_METRIC', item: newMetric, newGuesstimate: newGuesstimates[i]})})
+    if (existingMetrics.length > 0) {
+      dispatch(metricActions.removeMetrics(existingMetrics.map(m => m.id)))
+    }
+    if (newMetrics.length > 0) {
+      dispatch({type: 'ADD_METRICS', items: newMetrics, newGuesstimates: newGuesstimates})
+    }
 
     dispatch(runSimulations({spaceId, metricSubset: newMetrics}))
     dispatch(selectRegion(pasteRegion[0], pasteRegion[1]))
